keeps tabulation for vcp in main cli menu. moves cloudKeystore code to single file. refactors flags connection validations to be more specific for provision use case and in so, avoiding make any API call before reaching provision function
